requirements.txt
djangorestframework>=3.12.1,<3.13.0
pillow 10.2.0
